Further updates IMPORTANT COMMIT MESSAGE
We can get the very first adjacent next hexagon (at least for slope 1) and this works like a base case. Next we need to ensure GNA works or abandon it and use an alternative approach. Tomorrow, try to implement other functionalities first then come back to this at very latest.